printNumber

+++ 0x04-more_functions_nested_loops/main.h
@@ -12,7 +12,7 @@ void print_line(int n);
 void print_diagonal(int n);
 void print_square(int size);
 void print_triangle(int size);
-long long largestPrimeFactor(long long n);
+long largestPrimeFactor(long n);
 void print_number(int n);
 
 #endif
